backend/requirements.txt
django==4.0.3
Jinja2==3.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2==2.9.3
requests==2.28.1
pytz==2022.7.1
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1